eldercare/ictproject/voice-bot/requirements.txt
websocket-client==1.2.1
pyaudio==0.2.11
webrtcvad==2.0.10
numpy==1.21.0
requests==2.25.1